refactor(evals): move mintMcpToken/mcpAgentCall into the shared den-web lib
The agent-surface helpers were duplicated in member-scoped and exa flows
(plus an inline variant in cloud-oauth). All three now use the lib. Net
-60 lines; all three flows re-ran PASSED (note: on Chrome for Testing —
stable Chrome silently stopped honoring --disable-popup-blocking for
non-gesture window.open mid-day, breaking the OAuth-popup flows on any
stable Chrome; pinned CfT/Chromium is unaffected).
