+++ evals/flows/lib/den-web.mjs
@@ -100,3 +100,32 @@ export async function openYourConnections(ctx) {
   );
   await ctx.waitFor("window.location.pathname.endsWith('/your-connections')", { timeoutMs: 20_000, label: "Your Connections route" });
 }
+
+/** Mints an agent MCP token for a signed-in session (POST /v1/mcp/token). */
+export async function mintMcpToken(sessionToken, ctx) {
+  const { response, body } = await denApiFetch("/v1/mcp/token", {
+    method: "POST",
+    headers: { authorization: `Bearer ${sessionToken}` },
+    body: "{}",
+  });
+  ctx.assert(response.ok, `Minting an MCP token failed: ${response.status}`);
+  return body.token;
+}
+
+/** Calls the agent-facing MCP surface (/mcp/agent) and returns the JSON-RPC result. */
+export async function mcpAgentCall(mcpToken, method, params, ctx) {
+  const response = await fetch(`${denApiUrl()}/mcp/agent`, {
+    method: "POST",
+    headers: {
+      "content-type": "application/json",
+      accept: "application/json, text/event-stream",
+      authorization: `Bearer ${mcpToken}`,
+    },
+    body: JSON.stringify({ jsonrpc: "2.0", id: Date.now(), method, params }),
+  });
+  const raw = await response.text();
+  ctx.assert(response.ok, `MCP ${method} failed: ${response.status} ${raw.slice(0, 200)}`);
+  const dataLine = raw.split("\n").find((line) => line.startsWith("data:"));
+  ctx.assert(Boolean(dataLine), `MCP ${method} returned no data frame.`);
+  return JSON.parse(dataLine.slice(5)).result;
+}
