+++ evals/flows/mcp-connections-member-scoped.flow.mjs
@@ -36,7 +36,7 @@
  */
 
 import { execSync } from "node:child_process";
-import { denApiFetch, denApiUrl, openAdminConnections, openYourConnections, signInApi, signInViaBrowser } from "./lib/den-web.mjs";
+import { denApiFetch, mcpAgentCall, mintMcpToken, openAdminConnections, openYourConnections, signInApi, signInViaBrowser } from "./lib/den-web.mjs";
 
 const ADMIN_EMAIL = process.env.OPENWORK_EVAL_DEMO_EMAIL?.trim() || "alex@acme.test";
 const ADMIN_PASSWORD = process.env.OPENWORK_EVAL_DEMO_PASSWORD?.trim() || "OpenWorkDemo123!";
@@ -56,33 +56,6 @@ const state = {
   restrictedConnectionId: null,
   createdTeamId: null,
 };
-
-async function mcpAgentCall(mcpToken, method, params, ctx) {
-  const response = await fetch(`${denApiUrl()}/mcp/agent`, {
-    method: "POST",
-    headers: {
-      "content-type": "application/json",
-      accept: "application/json, text/event-stream",
-      authorization: `Bearer ${mcpToken}`,
-    },
-    body: JSON.stringify({ jsonrpc: "2.0", id: Date.now(), method, params }),
-  });
-  const raw = await response.text();
-  ctx.assert(response.ok, `MCP ${method} failed: ${response.status} ${raw.slice(0, 200)}`);
-  const dataLine = raw.split("\n").find((line) => line.startsWith("data:"));
-  ctx.assert(Boolean(dataLine), `MCP ${method} returned no data frame.`);
-  return JSON.parse(dataLine.slice(5)).result;
-}
-
-async function mintMcpToken(sessionToken, ctx) {
-  const { response, body } = await denApiFetch("/v1/mcp/token", {
-    method: "POST",
-    headers: { authorization: `Bearer ${sessionToken}` },
-    body: "{}",
-  });
-  ctx.assert(response.ok, `Minting an MCP token failed: ${response.status}`);
-  return body.token;
-}
 
 export default {
   id: "mcp-connections-member-scoped",
